Remove unnecessary memcpy from DDS loads
There's no postprocessing that needs to happen and there's no reason for the
headers to not stay in memory because they're small, so just grab the image data
by offsetting by the header size.

This saves a considerable amount of time loading Bistro. In Profile mode with
warm files being loaded, scene loading went from 1.93s (lowest) before to 1.37s
(lowest) after this change.

Platform::ReadEntireFile still takes the large majority of scene loading time.
As stated in the previous commit, this should be lowered by deduplicating loaded
images and parallelizing the load of Bistro's 600 files.