Add PBR workflow textures and use one global sampler
We need to support metallic-roughness and specular-glossiness because Sponza
uses the former while Bistro uses the later. There could be two separate
pipelines for each, but for now they're just the same shader with different
paths depending on the material buffer.

The sampler change is because the base GLTF standard doesn't support anisotropic
samplers and that's the only one I want to use right now anyways. The GLTF
parser still parses the samplers, the renderer just doesn't use them.