PBR rendering for metallic-roughness and specular-glossiness workflows
Note, this commit doesn't tonemap and therefore there is assumed to be only one
light which has a hardcoded radiance of 1.0.

Also, it might be worth it to compile two different pipelines, one for MR and
another for SG, to avoid the extra calculations and branches. Most scenes should
use one or the other either entirely or at least mostly. And either way, the
primitives could be sorted based on what pipeline they need to avoid unnecessary
pipeline switches.