Support shadows for transparent materials
Shadows are still very unoptimized, but here's some times from before this
change and now.

One directional light, 1080p, default camera position:
Sponza: 4.5ms -> 5.3ms
Bistro: 4.8ms -> 5.9ms

In terms of realism, shadows for blended materials could let in a fraction of
light dependent on the material's alpha. However, for now only the material's
alpha cutoff is used with no shadow factor accumulation. It's important to note
that translucent geometry which isn't alpha-tested is only visible on Bistro and
most of it is marked opaque while using the transmission GLTF extension.
Therefore, the effects of implementing this wouldn't be apparent yet.