Compute tangents in shader and support two channel normal maps
I exported Emerald Square and Sun Temple without tangents unlike Sponza and
Bistro which do have tangents. There's not much performance difference from
computing them as opposed to loading them.

Emerald Square and Sun Temple come with two channel BC5 compressed normal maps.
These are explicitly disallowed by GLTF's specification so I added an top-level
"extras" field so this option could still be data-driven. The saturate in the
normalMap.z calculation is due to some very rare NaNs that could be seen for
example on Emerald Square's center tree. It might be due to the BC5 encoding or
just an incorrect texture. The shadeNormal check is because a normalMap value of
(0.0, 0.0, 0.0) should encode no movement but will cause the geometry normal to
collapse. This could be seen on Emerald Square's lamp posts.