Fix shadows for mask materials
The issue with CommitNonOpaqueTriangleHit is that any hit further than the
committed hit will never be encountered. Combined with the fact that hit
traversal order isn't guaranteed to start from furthest and end with the
nearest, this means shadows can miss geometry. Now, hits are never committed to
avoid narrowing the search range. This is slightly less efficient, depending on
the amount of masked geometry, but acceptable for the great increase of shadow
quality.